outputting dynamic content from search results
using a template literal in javascript, I was able to accomplish mapping or cycling through the posts array from WP and outputting in a list item links to each title related to the searched text.  To eliminate the comma after each item in the array, I used join and set the parameter to an empty string.